+++ main.py
@@ -61,6 +61,7 @@ from src.backend.check_license import LicenseManager
 from src.frontend.UI.license import License, Disclaimer
 from src.backend.config import shared_config
 from src.backend.shared_functions import *
+
 from hqporner_api.api import Sort as hq_Sort
 from pathlib import Path